requirements.txt
arrow==1.2.3
binaryornot==0.4.4
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cookiecutter==2.1.1
cycler==0.11.0
Cython==0.29.28
docopt==0.6.2
fonttools==4.38.0
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
kivy-ios==2022.7.19
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
mouse==0.5.0
MouseInfo==0.1.3
numpy==1.23.4
openstep-parser==1.5.4
packaging==21.3
pbxproj==3.5.0
Pillow==9.2.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc==9.0
pyobjc-core==9.0
pyobjc-framework-Accessibility==9.0
pyobjc-framework-Accounts==9.0
pyobjc-framework-AddressBook==9.0
pyobjc-framework-AdServices==9.0
pyobjc-framework-AdSupport==9.0
pyobjc-framework-AppleScriptKit==9.0
pyobjc-framework-AppleScriptObjC==9.0
pyobjc-framework-ApplicationServices==9.0
pyobjc-framework-AppTrackingTransparency==9.0
pyobjc-framework-AudioVideoBridging==9.0
pyobjc-framework-AuthenticationServices==9.0
pyobjc-framework-AutomaticAssessmentConfiguration==9.0
pyobjc-framework-Automator==9.0
pyobjc-framework-AVFoundation==9.0
pyobjc-framework-AVKit==9.0
pyobjc-framework-AVRouting==9.0
pyobjc-framework-BackgroundAssets==9.0
pyobjc-framework-BusinessChat==9.0
pyobjc-framework-CalendarStore==9.0
pyobjc-framework-CallKit==9.0
pyobjc-framework-CFNetwork==9.0
pyobjc-framework-ClassKit==9.0
pyobjc-framework-CloudKit==9.0
pyobjc-framework-Cocoa==9.0
pyobjc-framework-Collaboration==9.0
pyobjc-framework-ColorSync==9.0
pyobjc-framework-Contacts==9.0
pyobjc-framework-ContactsUI==9.0
pyobjc-framework-CoreAudio==9.0
pyobjc-framework-CoreAudioKit==9.0
pyobjc-framework-CoreBluetooth==9.0
pyobjc-framework-CoreData==9.0
pyobjc-framework-CoreHaptics==9.0
pyobjc-framework-CoreLocation==9.0
pyobjc-framework-CoreMedia==9.0
pyobjc-framework-CoreMediaIO==9.0
pyobjc-framework-CoreMIDI==9.0
pyobjc-framework-CoreML==9.0
pyobjc-framework-CoreMotion==9.0
pyobjc-framework-CoreServices==9.0
pyobjc-framework-CoreSpotlight==9.0
pyobjc-framework-CoreText==9.0
pyobjc-framework-CoreWLAN==9.0
pyobjc-framework-CryptoTokenKit==9.0
pyobjc-framework-DataDetection==9.0
pyobjc-framework-DeviceCheck==9.0
pyobjc-framework-DictionaryServices==9.0
pyobjc-framework-DiscRecording==9.0
pyobjc-framework-DiscRecordingUI==9.0
pyobjc-framework-DiskArbitration==9.0
pyobjc-framework-DVDPlayback==9.0
pyobjc-framework-EventKit==9.0
pyobjc-framework-ExceptionHandling==9.0
pyobjc-framework-ExecutionPolicy==9.0
pyobjc-framework-ExtensionKit==9.0
pyobjc-framework-ExternalAccessory==9.0
pyobjc-framework-FileProvider==9.0
pyobjc-framework-FileProviderUI==9.0
pyobjc-framework-FinderSync==9.0
pyobjc-framework-FSEvents==9.0
pyobjc-framework-GameCenter==9.0
pyobjc-framework-GameController==9.0
pyobjc-framework-GameKit==9.0
pyobjc-framework-GameplayKit==9.0
pyobjc-framework-HealthKit==9.0
pyobjc-framework-ImageCaptureCore==9.0
pyobjc-framework-IMServicePlugIn==9.0
pyobjc-framework-InputMethodKit==9.0
pyobjc-framework-InstallerPlugins==9.0
pyobjc-framework-InstantMessage==9.0
pyobjc-framework-Intents==9.0
pyobjc-framework-IntentsUI==9.0
pyobjc-framework-IOSurface==9.0
pyobjc-framework-iTunesLibrary==9.0
pyobjc-framework-KernelManagement==9.0
pyobjc-framework-LatentSemanticMapping==9.0
pyobjc-framework-LaunchServices==9.0
pyobjc-framework-libdispatch==9.0
pyobjc-framework-LinkPresentation==9.0
pyobjc-framework-LocalAuthentication==9.0
pyobjc-framework-LocalAuthenticationEmbeddedUI==9.0
pyobjc-framework-MailKit==9.0
pyobjc-framework-MapKit==9.0
pyobjc-framework-MediaAccessibility==9.0
pyobjc-framework-MediaLibrary==9.0
pyobjc-framework-MediaPlayer==9.0
pyobjc-framework-MediaToolbox==9.0
pyobjc-framework-Metal==9.0
pyobjc-framework-MetalFX==9.0
pyobjc-framework-MetalKit==9.0
pyobjc-framework-MetalPerformanceShaders==9.0
pyobjc-framework-MetalPerformanceShadersGraph==9.0
pyobjc-framework-MetricKit==9.0
pyobjc-framework-MLCompute==9.0
pyobjc-framework-ModelIO==9.0
pyobjc-framework-MultipeerConnectivity==9.0
pyobjc-framework-NaturalLanguage==9.0
pyobjc-framework-NetFS==9.0
pyobjc-framework-Network==9.0
pyobjc-framework-NetworkExtension==9.0
pyobjc-framework-NotificationCenter==9.0
pyobjc-framework-OpenDirectory==9.0
pyobjc-framework-OSAKit==9.0
pyobjc-framework-OSLog==9.0
pyobjc-framework-PassKit==9.0
pyobjc-framework-PencilKit==9.0
pyobjc-framework-Photos==9.0
pyobjc-framework-PhotosUI==9.0
pyobjc-framework-PreferencePanes==9.0
pyobjc-framework-PushKit==9.0
pyobjc-framework-Quartz==9.0
pyobjc-framework-QuickLookThumbnailing==9.0
pyobjc-framework-ReplayKit==9.0
pyobjc-framework-SafariServices==9.0
pyobjc-framework-SafetyKit==9.0
pyobjc-framework-SceneKit==9.0
pyobjc-framework-ScreenCaptureKit==9.0
pyobjc-framework-ScreenSaver==9.0
pyobjc-framework-ScreenTime==9.0
pyobjc-framework-ScriptingBridge==9.0
pyobjc-framework-SearchKit==9.0
pyobjc-framework-Security==9.0
pyobjc-framework-SecurityFoundation==9.0
pyobjc-framework-SecurityInterface==9.0
pyobjc-framework-ServiceManagement==9.0
pyobjc-framework-SharedWithYou==9.0
pyobjc-framework-SharedWithYouCore==9.0
pyobjc-framework-ShazamKit==9.0
pyobjc-framework-Social==9.0
pyobjc-framework-SoundAnalysis==9.0
pyobjc-framework-Speech==9.0
pyobjc-framework-SpriteKit==9.0
pyobjc-framework-StoreKit==9.0
pyobjc-framework-SyncServices==9.0
pyobjc-framework-SystemConfiguration==9.0
pyobjc-framework-SystemExtensions==9.0
pyobjc-framework-ThreadNetwork==9.0
pyobjc-framework-UniformTypeIdentifiers==9.0
pyobjc-framework-UserNotifications==9.0
pyobjc-framework-UserNotificationsUI==9.0
pyobjc-framework-VideoSubscriberAccount==9.0
pyobjc-framework-VideoToolbox==9.0
pyobjc-framework-Virtualization==9.0
pyobjc-framework-Vision==9.0
pyobjc-framework-WebKit==9.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
python-slugify==6.1.2
pytweening==1.0.4
PyYAML==6.0
requests==2.28.1
rubicon-objc==0.4.2
sh==1.14.3
six==1.16.0
text-unidecode==1.3
tkthread==0.5.2
urllib3==1.26.12
